requirements.txt
# Core
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.6.0
pydantic-settings>=2.2.0

# LLM & Embeddings
httpx>=0.27.0
langchain>=0.2.0
langchain-community>=0.2.0
langchain-text-splitters>=0.2.0

# Vector Stores
chromadb>=0.5.0
faiss-cpu>=1.8.0

# Document Loaders
pypdf>=4.0.0
docx2txt>=0.8
unstructured>=0.13.0

# Reranking
sentence-transformers>=3.0.0

# UI
streamlit>=1.35.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
